Refactor scenes into different packages
