+++ src/Main/Main.java
@@ -11,10 +11,10 @@ public class Main extends Application {
     @Override
     public void start(Stage primaryStage) throws Exception{
         Parent root = FXMLLoader.load(getClass().getResource("Main.fxml"));
-        Parent root1 = FXMLLoader.load(getClass().getResource("Podcast.fxml"));
-        Parent root2 = FXMLLoader.load(getClass().getResource("Search.fxml"));
+        Parent root1 = FXMLLoader.load(getClass().getResource("../Podcast/Podcast.fxml"));
+        Parent root2 = FXMLLoader.load(getClass().getResource("../Search/Search.fxml"));
         primaryStage.setTitle("Hello World");
-        primaryStage.setScene(new Scene(root, 1280, 720));
+        primaryStage.setScene(new Scene(root2, 1280, 720));
         primaryStage.show();
     }
