--- src/Main/Podcast.fxml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.control.ToolBar?>
-<?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.text.Text?>
-
-
-<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
-   <top>
-      <ToolBar prefHeight="40.0" prefWidth="200.0" BorderPane.alignment="CENTER">
-        <items>
-            <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Your Podcasts" />
-            <TextField translateX="300.0" />
-          <Button mnemonicParsing="false" text="Search" translateX="300.0" />
-        </items>
-      </ToolBar>
-   </top>
-</BorderPane>
