+++ src/Podcast/Podcast.fxml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.scene.text.*?>
+
+<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0"
+            prefWidth="600.0" xmlns="http://javafx.com/javafx/10.0.2-internal">
+   <top>
+       <ToolBar prefHeight="40.0" prefWidth="200.0" BorderPane.alignment="CENTER">
+           <Text strokeType="OUTSIDE" strokeWidth="0.0" text="Your Podcasts"/>
+           <TextField alignment="TOP_LEFT" translateX="300.0"/>
+           <Button alignment="TOP_LEFT" mnemonicParsing="false" text="Search" translateX="300.0"/>
+       </ToolBar>
+   </top>
+</BorderPane>
